+++ src/database/operations/changemakerFieldValues/index.ts
@@ -1,1 +1,3 @@
 export * from './createChangemakerFieldValue';
+export * from './loadChangemakerFieldValue';
+export * from './loadChangemakerFieldValueBundle';