+++ src/routers/changemakerFieldValuesRouter.ts
@@ -4,6 +4,18 @@ import { requireAuthentication } from '../middleware';
 
 const changemakerFieldValuesRouter = express.Router();
 
+changemakerFieldValuesRouter.get(
+	'/',
+	requireAuthentication,
+	changemakerFieldValuesHandlers.getChangemakerFieldValues,
+);
+
+changemakerFieldValuesRouter.get(
+	'/:fieldValueId',
+	requireAuthentication,
+	changemakerFieldValuesHandlers.getChangemakerFieldValue,
+);
+
 changemakerFieldValuesRouter.post(
 	'/',
 	requireAuthentication,